refactor(edge): clean up control plane module boundaries
